+++ src/lib.rs
@@ -67,15 +67,10 @@
 // #![warn(missing_docs)]
 #![allow(clippy::missing_safety_doc)]
 
-#[macro_use]
-extern crate educe;
-
 pub use self::{
     compiler::compile,
-    jit_runner::{execute, execute_contract},
+    ffi::{module_to_object, object_to_shared_lib, LLVMCompileError},
 };
-
-pub use self::jit_runner::ExecutionResult;
 
 pub mod cache;
 mod compiler;
@@ -85,7 +80,7 @@ pub mod error;
 pub mod execution_result;
 pub mod executor;
 mod ffi;
-mod jit_runner;
+// mod jit_runner;
 pub mod libfuncs;
 pub mod metadata;
 pub mod module;
@@ -93,4 +88,3 @@ pub mod starknet;
 pub mod types;
 pub mod utils;
 pub mod values;
-pub use ffi::{module_to_object, object_to_shared_lib, LLVMCompileError};
